requirements.txt
numpy
pandas
sqlalchemy
bokeh
ipykernel
notebook
pyyaml